rename scss files
