+++ packages/gamut/src/AppBar/styles/index.module.scss
